+++ src/details/ArborX_DetailsBatchedQueries.hpp
@@ -171,11 +171,11 @@ public:
 
   static std::tuple<Kokkos::View<int *, DeviceType>,
                     Kokkos::View<int *, DeviceType>,
-                    Kokkos::View<double *, DeviceType>>
+                    Kokkos::View<float *, DeviceType>>
   reversePermutation(Kokkos::View<size_t const *, DeviceType> permute,
                      Kokkos::View<int const *, DeviceType> offset,
                      Kokkos::View<int const *, DeviceType> indices,
-                     Kokkos::View<double const *, DeviceType> distances)
+                     Kokkos::View<float const *, DeviceType> distances)
   {
     auto const tmp_offset = permuteOffset(permute, offset);
